FIAP/Cap 1 - FarmTech na era da cloud computing/requirements.txt
pandas>=1.3
numpy>=1.21
scikit-learn>=0.24
xgboost>=1.5
joblib>=1.1
flask>=2.0
gunicorn>=20.1
paho-mqtt>=1.6
matplotlib>=3.4
seaborn>=0.11
sqlalchemy>=1.4
python-dotenv>=0.19
requests>=2.26